refactor: update component imports and exports for consistency, remove unused index file
