+++ src/components/BackgroundInformationView/index.ts
@@ -1,0 +1,1 @@
+export { BackgroundInformationView } from './BackgroundInformationView'
