+++ src/components/BaseMapPopup/index.ts
@@ -1,0 +1,1 @@
+export { BaseMapPopup } from './BaseMapPopup'
