+++ src/components/BuildingsCard/BuildingsCard.tsx
@@ -1,5 +1,5 @@
 import { Card, Flex, Text } from '@radix-ui/themes'
-import { CustomPopover } from '../CustomPopover/CustomPopover'
+import { CustomPopover } from '../CustomPopover'
 
 export const BuildingsCard = ({ numberOfBuildings }: { numberOfBuildings: number | string }) => {
   return (
