+++ src/components/BuildingsCard/index.ts
@@ -1,0 +1,1 @@
+export { BuildingsCard } from './BuildingsCard'
