+++ src/components/ChartCard/index.ts
@@ -1,0 +1,1 @@
+export { ChartCard } from './ChartCard'
