+++ src/components/CountryResultView/CountryResultView.tsx
@@ -1,12 +1,12 @@
 import useResponsive from '../../hooks/useResponsive'
 import { Flex, Grid } from '@radix-ui/themes'
-import { ShorelineChangeCard } from '../ShoreLineChangeCard/ShoreLineChangeCard'
-import { HotSpotsCard } from '../HotSpotsCard/HotSpotsCard'
-import { PopulationCard } from '../PopulationCard/PopulationCard'
-import { BuildingsCard } from '../BuildingsCard/BuildingsCard'
-import { MangrovesCard } from '../MangrovesCard/MangrovesCard'
-import { ChartCard } from '../ChartCard/ChartCard'
-import TextButton from '../TextButton/TextButton'
+import { ShorelineChangeCard } from '../ShoreLineChangeCard'
+import { HotSpotsCard } from '../HotSpotsCard'
+import { PopulationCard } from '../PopulationCard'
+import { BuildingsCard } from '../BuildingsCard'
+import { MangrovesCard } from '../MangrovesCard'
+import { ChartCard } from '../ChartCard'
+import { TextButton } from '../TextButton'
 import type { ContiguousHotspotProperties } from '../../library/types/countryGeoJsonTypes'
 import { useMapData } from '../../hooks/useGlobalContext'
