+++ src/components/CountryResultView/index.ts
@@ -1,0 +1,1 @@
+export { CountryResultView } from './CountryResultView'
