+++ src/components/CustomPopover/index.ts
@@ -1,0 +1,1 @@
+export { CustomPopover } from './CustomPopover'
