+++ src/components/HotSpotsCard/index.ts
@@ -1,0 +1,1 @@
+export { HotSpotsCard } from './HotSpotsCard'
