+++ src/components/MangrovesCard/MangrovesCard.tsx
@@ -1,5 +1,5 @@
 import { Card, Flex, Text } from '@radix-ui/themes'
-import { CustomPopover } from '../CustomPopover/CustomPopover'
+import { CustomPopover } from '../CustomPopover'
 
 export const MangrovesCard = ({ mangroveArea }: { mangroveArea: number | string }) => {
   return (
