+++ src/components/MangrovesCard/index.ts
@@ -1,0 +1,1 @@
+export { MangrovesCard } from './MangrovesCard'
