+++ src/components/PopulationCard/index.ts
@@ -1,0 +1,1 @@
+export { PopulationCard } from './PopulationCard'
