+++ src/components/RateOfChangeCard/RateOfChangeCard.tsx
@@ -1,5 +1,5 @@
 import { Card, Flex, Text } from '@radix-ui/themes'
-import { CustomPopover } from '../CustomPopover/CustomPopover'
+import { CustomPopover } from '../CustomPopover'
 
 export const RateOfChangeCard = ({ rateOfChange }: { rateOfChange: number | string }) => {
   return (
