+++ src/components/RateOfChangeCard/index.ts
@@ -1,0 +1,1 @@
+export { RateOfChangeCard } from './RateOfChangeCard'
