+++ src/components/ResultPanel/ResultPanel.tsx
@@ -3,11 +3,11 @@ import { useState, useEffect } from 'react'
 import { MobileResultBottomPanel } from '../MobileResultBottomPanel'
 import useResponsive from '../../hooks/useResponsive'
 import styles from './ResultPanel.module.scss'
-import { ErrorCard } from '../ErrorCard/ErrorCard'
-import { CountryResultView } from '../CountryResultView/CountryResultView'
-import { LocationCard } from '../LocationCard/LocationCard'
-import { HotSpotResultView } from '../HotSpotResultView/HotSpotResultView'
-import BackgroundInformationView from '../BackgroundInformationView/BackgroundInformationView'
+import { ErrorCard } from '../ErrorCard'
+import { CountryResultView } from '../CountryResultView'
+import { LocationCard } from '../LocationCard'
+import { HotSpotResultView } from '../HotSpotResultView'
+import { BackgroundInformationView } from '../BackgroundInformationView'
 import { useMapData } from '../../hooks/useGlobalContext'
 import type { ContiguousHotspotProperties } from '../../library/types/countryGeoJsonTypes'
