+++ src/components/TextButton/index.ts
@@ -1,0 +1,1 @@
+export { TextButton } from './TextButton'
