--- src/components/index.ts
@@ -1,3 +1,0 @@
-export { Dashboard } from './Dashboard'
-export { MainMap } from './MainMap'
-export { ResultPanel } from './ResultPanel'
